01_30_2009 PC
git-svn-id: http://dim3.googlecode.com/svn/trunk@394 2c4a6428-2a2a-0410-ad13-7dde452bfe2e

+++ dim3MapUtility/Sources/map_mesh_update.c
@@ -1071,8 +1071,8 @@ void map_mesh_whole_poly_uv(map_type *map,int mesh_idx,int poly_idx)
 	poly=&mesh->polys[poly_idx];
 
 	for (n=0;n!=poly->ptsz;n++) {
-		poly->gx[n]=floor(poly->gx[n]);
-		poly->gy[n]=floor(poly->gy[n]);
+		poly->gx[n]=(float)floor(poly->gx[n]);
+		poly->gy[n]=(float)floor(poly->gy[n]);
 	}
 }
